1.2.2

+++ texts/1章/main.py
@@ -15,7 +15,7 @@ print(array4.dtype)
 array3[0,2] = 0.5j
 print(array3)
 
-array1 = np.empty(10, dtype=float)
+array1 = np.empty(10, dtype=int)
 print(array1)
 
 array1 = np.array([[0, -1j], [1j, 0]], dtype=complex)
